Modify score to float, and add GET (#22)
* Modify score to float, and add GET

* Save to DB as integer

* Refactor score endpoints

* Extract scores endpoints to separate router

* Rename tags

+++ src/api/model/training.py
@@ -132,11 +132,11 @@ class BlockStatus(BaseModel):
 
 
 class ScoreBody(BaseModel):
-    score: int = Field(
+    score: float = Field(
         title="Score",
         description="Score given to training by user",
-        ge=0,
-        le=5,
+        ge=0.0,
+        le=5.0,
     )
